+++ lib/pages/location_list_page.dart
@@ -2,6 +2,7 @@ import 'package:belleza_app/database/database_helper.dart';
 import 'package:belleza_app/pages/add_location_page.dart';
 import 'package:belleza_app/pages/edit_location_page.dart';
 import 'package:belleza_app/pages/location_products_page.dart';
+import 'package:belleza_app/utils/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 
@@ -13,7 +14,6 @@ class LocationListPage extends StatefulWidget {
 }
 
 class _LocationListPageState extends State<LocationListPage> {
-
   List<Map<String, dynamic>> _locations = [];
 
   final dbHelper = DatabaseHelper();
@@ -30,14 +30,23 @@ class _LocationListPageState extends State<LocationListPage> {
       _locations = locations;
     });
   }
+
   void _deleteLocation(int id) async {
-    await dbHelper.deleteLocation(id);
-    _loadLocations(); // Recargar la lista de proveedores después de eliminar
+    final confirmed = await Utils.showConfirmationDialog(
+      context,
+      'Confirmar eliminación',
+      '¿Estás seguro de que deseas eliminar esta ubicación?',
+    );
+    if (confirmed) {
+      await dbHelper.deleteLocation(id);
+      _loadLocations();
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
+      backgroundColor: Utils.colorFondo,
       body: ListView.builder(
         itemCount: _locations.length,
         itemBuilder: (context, index) {
@@ -50,7 +59,7 @@ class _LocationListPageState extends State<LocationListPage> {
               ));
             },
             child: Card(
-              color: Colors.pink[100], // Color de fondo del Card
+              color: Utils.colorFondoCards, // Color de fondo del Card
               margin: const EdgeInsets.symmetric(
                   vertical: 8.0, horizontal: 12.0), // Márgenes del Card
               shape: RoundedRectangleBorder(
@@ -58,65 +67,65 @@ class _LocationListPageState extends State<LocationListPage> {
               ),
               elevation: 4, // Sombra para el Card
               child: Padding(
-                padding: const EdgeInsets.only(left: 12.0), // Espaciado interno del Card
+                padding: const EdgeInsets.only(
+                    left: 12.0), // Espaciado interno del Card
                 child: Row(
                   children: [
                     Expanded(
                       child: Column(
                         children: [
                           Text(
-                        location['name'] ?? 'Sin datos',
-                        style: const TextStyle(
-                          fontSize: 18.0,
-                          fontWeight: FontWeight.bold,
-                        ),
-                      ),
-                      const SizedBox(height: 8.0),
-                      Text(
-                        location['description'] ?? 'Sin datos',
-                        style: const TextStyle(
-                          fontSize: 14.0,
-                          color: Colors.black87,
-                        ),
-                      ),
+                            location['name'] ?? 'Sin datos',
+                            style: const TextStyle(
+                              fontSize: 18.0,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                          const SizedBox(height: 8.0),
+                          Text(
+                            location['description'] ?? 'Sin datos',
+                            style: const TextStyle(
+                              fontSize: 14.0,
+                              color: Colors.black87,
+                            ),
+                          ),
                         ],
                       ),
                     ),
                     Column(
-                          mainAxisSize: MainAxisSize.min,
-                          children: [
-                            Container(
-                              decoration: BoxDecoration(
-                                color: Colors.blue, // Color de fondo
-                                borderRadius: BorderRadius.only(
-                                    topRight:
-                                        Radius.circular(16)), // Bordes redondeados
-                              ), // Espaciado interno
-                              child: IconButton(
-                                onPressed: () {
-                                  Get.to(EditLocationPage(location: location));
-                                },
-                                icon: const Icon(Icons.edit),
-                                color: Colors.white, // Color del ícono
-                              ),
-                            ),
-                            Container(
-                              decoration: BoxDecoration(
-                                color: Colors.red,
-                                borderRadius: BorderRadius.only(
-                                    bottomRight:
-                                        Radius.circular(16)),
-                              ),
-                              child: IconButton(
-                                onPressed: () {
-                                  _deleteLocation(location['id']);
-                                },
-                                icon: const Icon(Icons.delete),
-                                color: Colors.white, // Color del ícono
-                              ),
-                            ),
-                          ],
+                      mainAxisSize: MainAxisSize.min,
+                      children: [
+                        Container(
+                          decoration: BoxDecoration(
+                            color: Utils.edit, // Color de fondo
+                            borderRadius: BorderRadius.only(
+                                topRight:
+                                    Radius.circular(16)), // Bordes redondeados
+                          ), // Espaciado interno
+                          child: IconButton(
+                            onPressed: () {
+                              Get.to(EditLocationPage(location: location));
+                            },
+                            icon: const Icon(Icons.edit),
+                            color: Colors.white, // Color del ícono
+                          ),
                         ),
+                        Container(
+                          decoration: BoxDecoration(
+                            color: Utils.delete,
+                            borderRadius: BorderRadius.only(
+                                bottomRight: Radius.circular(16)),
+                          ),
+                          child: IconButton(
+                            onPressed: () {
+                              _deleteLocation(location['id']);
+                            },
+                            icon: const Icon(Icons.delete),
+                            color: Colors.white, // Color del ícono
+                          ),
+                        ),
+                      ],
+                    ),
                   ],
                 ),
               ),
@@ -128,9 +137,11 @@ class _LocationListPageState extends State<LocationListPage> {
         backgroundColor: Colors.pink,
         onPressed: () async {
           Get.to(AddLocationPage());
-        
         },
-        child: const Icon(Icons.add, color: Colors.white,),
+        child: const Icon(
+          Icons.add,
+          color: Colors.white,
+        ),
       ),
     );
   }
